use bitwise ops rather than B to check for numeric flags
This abuses the fact that bitwise ops change behavior based on the
internal type of a value.  This avoids the memory cost of loading B, and
ends up being faster as well.

We first us an & to shorten the input, so we don't end up operating on a
long string.

"0" & 0       -> 0
"0" & ""      -> ""
"0" & $string -> $character

0 ^ 0                    -> 0 (false)
$character ^ $character  -> "\x00" (true)